Put utils files in separate folder
