+++ src/export.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { QuickPickItem, window } from 'vscode';
-import { MultiStepInput, toQuickPickItem, toQuickPickItems, shouldResumeNoOp } from './multiStepVscode';
+import { MultiStepInput, toQuickPickItem, toQuickPickItems, shouldResumeNoOp } from './utils/multiStepQuickPickUtil';
 import { findClassLines } from './pythonParsing';
 import { Logger, Window } from './logger';
 import { waitNewTerminalDelay } from './manimShell';
